putting all the document functionalities together
