+++ webapi/appsettings.json
@@ -38,7 +38,7 @@
   //
   "AIService": {
     "Type": "AzureOpenAI",
-    "Endpoint": "", // ignored when AIService is "OpenAI"
+    "Endpoint": "https://teams-bot.openai.azure.com/", // ignored when AIService is "OpenAI"
     // "Key": "",
     "Models": {
       "Completion": "gpt-35-turbo", // For OpenAI, change to 'gpt-3.5-turbo' (with a period).
@@ -233,4 +233,3 @@
   //
   "APPLICATIONINSIGHTS_CONNECTION_STRING": null
 }
-
