+++ webapp/src/libs/services/DocumentImportService.ts
@@ -29,4 +29,25 @@ export class DocumentImportService extends BaseService {
             accessToken,
         );
     };
+
+    public deleteDocumentAsync = async (
+        userId: string,
+        chatId: string,
+        documentId: string,
+        accessToken: string,
+    ) => {
+        const formData = new FormData();
+        formData.append('userId', userId);
+        formData.append('chatId', chatId);
+        formData.append('documentId', documentId);
+
+        return await this.getResponseAsync<string>(
+            {
+                commandPath: 'deleteDocument',
+                method: 'POST',
+                body: formData,
+            },
+            accessToken,
+        );
+    };
 }
